feat(profile): Add profile and change password
