+++ src/features/account/pages/ProfilePage.tsx
@@ -1,0 +1,177 @@
+import React, { useState, useEffect } from "react";
+import { 
+  Typography, 
+  Box, 
+  Fade, 
+  Alert,
+  Skeleton,
+  Card,
+  CardContent
+} from "@mui/material";
+import { 
+  ErrorOutline, 
+  
+  
+} from "@mui/icons-material";
+import { useAuth } from "../../auth/hooks/authContext";
+import { useProfile } from "../../auth/hooks/useprofile";
+import useChangePassword from "../hooks/useChangePassword";
+import { useNotification } from "../../../contexts/NotificationContext";
+import ProfileLayout from "../components/ProfileLayout";
+import AccountInfo from "../components/AccountInfo";
+import ChangePasswordForm from "../components/ChangePasswordForm";
+import type { ChangePasswordData } from "../types/ChangePasswordForm.types";
+
+const ProfilePage: React.FC = () => {
+  const { user, loading: authLoading } = useAuth();
+  const { isFetching } = useProfile();
+  const { mutate: changePassword, status: changePasswordStatus } = useChangePassword();
+  const { showSuccessNotification, showErrorNotification } = useNotification();
+  const [error, setError] = useState("");
+  const [mounted, setMounted] = useState(false);
+
+  // Set document title and mounted state when component mounts
+  useEffect(() => {
+    document.title = "Thông tin tài khoản";
+    setMounted(true);
+    
+    // Add a subtle page transition
+    const timer = setTimeout(() => {
+      setMounted(true);
+        }, 100);
+        
+    return () => clearTimeout(timer);
+  }, []);
+
+  const handleChangePassword = (formData: ChangePasswordData) => {
+    setError("");
+    changePassword(
+      {
+        currentPassword: formData.currentPassword,
+        newPassword: formData.newPassword,
+        confirmNewPassword: formData.confirmNewPassword,
+      },
+      {
+        onSuccess: () => {
+          showSuccessNotification("Đổi mật khẩu thành công! Vui lòng đăng nhập lại.");
+          
+          // Add a delay before redirecting for better UX
+          setTimeout(() => {
+            document.cookie = "feAccessToken=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;";
+            window.location.href = "/login";
+          }, 1500);
+        },
+        onError: (error: any) => {
+          const apiMessage = error.response?.data?.message;
+          const errorMsg = apiMessage || "Đổi mật khẩu thất bại! Vui lòng thử lại.";
+          showErrorNotification(errorMsg);
+          setError(errorMsg);
+        },
+      }
+    );
+  };
+
+  // Loading skeleton component
+  const LoadingSkeleton = () => (
+    <Box sx={{ space: 3 }}>
+      <Card sx={{ mb: 4, borderRadius: '16px' }}>
+        <CardContent sx={{ p: 4 }}>
+          <Box sx={{ display: 'flex', alignItems: 'center', mb: 3 }}>
+            <Skeleton variant="circular" width={64} height={64} sx={{ mr: 3 }} />
+            <Box sx={{ flex: 1 }}>
+              <Skeleton variant="text" width="60%" height={32} sx={{ mb: 1 }} />
+              <Box sx={{ display: 'flex', gap: 1 }}>
+                <Skeleton variant="rounded" width={80} height={24} />
+                <Skeleton variant="rounded" width={100} height={24} />
+              </Box>
+            </Box>
+          </Box>
+          <Skeleton variant="text" width="40%" height={24} sx={{ mb: 2 }} />
+          <Skeleton variant="rounded" width="100%" height={56} sx={{ mb: 2 }} />
+          <Skeleton variant="rounded" width="100%" height={56} />
+        </CardContent>
+      </Card>
+      
+      <Card sx={{ borderRadius: '16px' }}>
+        <CardContent sx={{ p: 4 }}>
+          <Skeleton variant="text" width="30%" height={24} sx={{ mb: 3 }} />
+          <Skeleton variant="rounded" width="100%" height={56} sx={{ mb: 2 }} />
+          <Skeleton variant="rounded" width="100%" height={56} sx={{ mb: 2 }} />
+          <Skeleton variant="rounded" width="100%" height={56} sx={{ mb: 3 }} />
+          <Skeleton variant="rounded" width="100%" height={48} />
+        </CardContent>
+      </Card>
+    </Box>
+  );
+
+  // Error state component
+  const ErrorState = () => (
+    <Fade in timeout={600}>
+      <Alert 
+        severity="error" 
+        icon={<ErrorOutline />}
+        sx={{
+          borderRadius: '16px',
+          border: '1px solid',
+          borderColor: 'error.light',
+          p: 3,
+          '& .MuiAlert-icon': {
+            fontSize: '28px'
+          }
+        }}
+      >
+        <Typography variant="h6" sx={{ fontWeight: 600, mb: 1 }}>
+          Không thể tải thông tin tài khoản
+        </Typography>
+        <Typography variant="body2">
+          Vui lòng đăng nhập lại để tiếp tục sử dụng dịch vụ.
+        </Typography>
+      </Alert>
+    </Fade>
+  );
+
+  
+
+  return (
+    <ProfileLayout>
+      <Fade in={mounted} timeout={800}>
+        <Box>
+          {authLoading || isFetching ? (
+            <LoadingSkeleton />
+          ) : user ? (
+            <Box sx={{ 
+              '& > *': { 
+                opacity: 0,
+                animation: 'slideInUp 0.6s ease-out forwards'
+              },
+              '& > *:nth-of-type(2)': {
+                animationDelay: '0.2s'
+              },
+              '@keyframes slideInUp': {
+                from: {
+                  opacity: 0,
+                  transform: 'translateY(20px)'
+                },
+                to: {
+                  opacity: 1,
+                  transform: 'translateY(0)'
+                }
+              }
+            }}>
+              <AccountInfo user={user} />
+              <ChangePasswordForm
+                onSubmit={handleChangePassword}
+                error={error}
+                changePasswordStatus={changePasswordStatus}
+              />
+            </Box>
+          ) : (
+            <ErrorState />
+          )}
+        </Box>
+      </Fade>
+    </ProfileLayout>
+  );
+};
+
+export default ProfilePage;
