+++ src/features/account/types/AccountInfo.types.ts
@@ -1,0 +1,5 @@
+import type { User } from "../types/auth.types";
+
+export interface AccountInfoProps {
+  user: User;
+}
