+++ src/features/account/types/ChangePasswordForm.types.ts
@@ -1,0 +1,11 @@
+export interface ChangePasswordData {
+  currentPassword: string;
+  newPassword: string;
+  confirmNewPassword: string;
+}
+
+export interface ChangePasswordFormProps {
+  onSubmit: (formData: ChangePasswordData) => void;
+  error: string;
+  changePasswordStatus: string;
+}
