+++ src/features/account/types/ProfileLayout.types.ts
@@ -1,0 +1,3 @@
+export interface ProfileLayoutProps {
+  children: React.ReactNode;
+}
