+++ src/features/account/types/auth.types.ts
@@ -1,0 +1,6 @@
+export interface User {
+  username: string;
+  email: string;
+  role: string;
+  isActive: boolean;
+}
